{"schema":"decodex/commit/1","summary":"Extract capture host view","authority":"manual"}
